Xcode 4.4
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@10136 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCM Framework/DCMAttribute.m
@@ -485,8 +485,8 @@
 
 - (NSString *)description{
 	if (self.valueLength < 100)
-		return  [NSString stringWithFormat:@"%@\t %@\t vl:%d\t vm:%d\t %@", _tag.description, _tag.vr, self.valueLength, self.valueMultiplicity, [self valuesAsString]];
-	return  [NSString stringWithFormat:@"%@\t vl:%d\t vm:%d", _tag.description, self.valueLength, self.valueMultiplicity];
+		return  [NSString stringWithFormat:@"%@\t %@\t vl:%d\t vm:%d\t %@", _tag.description, _tag.vr, (int)self.valueLength, self.valueMultiplicity, [self valuesAsString]];
+	return  [NSString stringWithFormat:@"%@\t vl:%d\t vm:%d", _tag.description, (int) self.valueLength, self.valueMultiplicity];
 }
