+++ Wait.m
@@ -92,7 +92,7 @@
 				seconds = fullWork;
 				seconds -= hours*60*60 + minutes*60;
 				
-				[elapsed setStringValue:[NSString stringWithFormat:@"Estimated remaining time: %2.2d:%2.2d:%2.2d", hours, minutes, seconds]];
+				[elapsed setStringValue:[NSString stringWithFormat:@"Estimated remaining time: %2.2d:%2.2d:%2.2d", (int) hours, (int) minutes, (int) seconds]];
 				[elapsed displayIfNeeded];
 				
 				#if __LP64__
